[CO-296] Extract and make ZenDesk API testable.
