+++ src/CLI.hs
@@ -7,7 +7,7 @@ import           Universum
 
 import           Options.Applicative (Parser, ParserInfo, argument, auto, command, execParser, fullDesc, header,
                                       help, helper, hsubparser, info, infoOption, long, metavar,
-                                      progDesc, strOption, (<**>))
+                                      progDesc, (<**>))
 import           Paths_log_classifier (version)
 
 data CLI
@@ -15,7 +15,6 @@ data CLI
     | ProcessTicket Int -- ^ Process ticket of an given ticket id
     | ProcessTickets    -- ^ Process all the tickets in Zendesk
     | FetchTickets      -- ^ Fetch all the tickets in Zendesk
-    | RawRequest String -- ^ Raw request to the given url
     | ShowStatistics    -- ^ Show statistics
     deriving (Show)
 
@@ -24,12 +23,6 @@ cmdProcessTicket :: Parser CLI
 cmdProcessTicket = ProcessTicket <$> argument auto
                        (metavar "TICKET_ID"
                        <> help "Ticket id to analyze")
-
--- | Parser for RawRequest
-cmdRawRequest :: Parser CLI
-cmdRawRequest = RawRequest <$> strOption
-                    (metavar "URL"
-                    <> help "Url to request")
 
 -- | Parser for CLI commands
 cli :: Parser CLI
@@ -42,8 +35,6 @@ cli = hsubparser $ mconcat
         (progDesc "Fetch all the tickets that need to be analyzes."))
     , command "process-ticket" (info cmdProcessTicket
         (progDesc "Process Zendesk ticket of an given ticket id"))
-    , command "raw-request" (info cmdRawRequest
-        (progDesc "Raw request to the given url"))
     , command "show-stats" (info (pure ShowStatistics)
         (progDesc "Print list of ticket Ids that agent has been assigned"))
     ]
